feat: organize the letsql package
The python package is composed of the following packages:

backends
code for the ibis backends, includes LetSQL (let) and DataFusion (datafusion).
More backends could be add in the future (extensible)

common
code for misc tasks, like the deferred in-disk cache

expr
code for both ibis and datafusion expr, and the transformation between them

optimizer
code for exposing the datafusion optimizer, also include functions for optimizing
ibis expr. The idea is to let the user of the library easily add their own
optimization rules

sql
code for exposing the datafusion parser and planner

The organization is heavily influenced by the one found in datafusion and ibis
(with a side of sqlglot), with this I'm hoping to keep it familiar with
the user base of those libraries.