feat: drop connection handling
Since LetSQL now has the capacity for adding ibis.Table as a TableProvider
we can drop the connection management, this is good because:

- it simplifies the code, no need to handle the connection directly
- from the UX point of view the user doesn't need to add a static reference
- The user doesn't need to learn a new set of methods

Now every time a table is registered, we add the original backend to keep
track of it, to execute single-backend queries on the original
source.

As of now directly executing on the original backend generates issues
when the table name changes (the test_register_already_existing_table fails).
So the execution goes through the to_pyarrow_batches method.

Additionally this commit removes unused code and parameters.