fix: allow for multiple execution of letsql tables
Why was a letsql table being executed only once?

Because when executing the TableProvider for IbisTable it would raise:

Cannot start a runtime from within a runtime. This happens because a function
(like block_on) attempted to block the current thread while the thread is being
used to drive asynchronous tasks.

Why the current code implemented?

Because in contrast to IbisTable, the TableProvider for PyRecordBatchProvider
would not raise an exception, but it had the downside of only being able to
execute it once.

Note there was a TODO to improve it in the future:

> make a better fix to make the table durable (like a datafusion view)

Why the IbisTable raised an exception and the PyRecordBatchProvider not?

The best guess is that the IbisTable calls to_pyarrow_batches, so there was
a nested call to_pyarrow_batches that created a runtime within a runtime.

Why the current solution works?
Because it allows to create a runtime within a runtime, see this post:
https://medium.com/@greptime_team/bridging-async-and-sync-rust-code-a
-lesson-learned-while-working-with-tokio-6173b1efaca4

and also the discussion herein:
https://www.reddit.com/r/rust/comments/11nyax9/bridging_async_and_sync
_rust_code_a_lesson/

closes #39