fix: change name of parameter in replace_table function
The name n makes it clash with operations/expressions containing
arguments with that name, for example limit.

The root of the issue is that the map function creates a kwargs
dictionary containing as keys the arguments of the operation.

When the name of one of the arguments of the operation match the
name of the first argument the issue arises.