chore: set better configuration for letsql package
The file was moved to inside the .github directory, renovate should be able to
find it, see this [resource](https://docs.renovatebot.com/configuration-options/)

rangeStrategy is set to "widen"

> widen = Widen the range with newer one, e.g. ^1.0.0 -> ^1.0.0 || ^2.0.0

it will also widen other types of ranges, e.g.

>=13.0.0,<14 -> >=13.0.0,<18

semanticCommits was enabled to force Semantic Commit prefixes for commits and PR titles.

lockFileMaintenance was enabled //TODO this probably needs review (at least for cargo)

add the manager pep261 (is different for the poetry) to keep in update the depencies under project TOML table

autoMerge was set to false, we only merge after reviewing the PR and approving it

printConfig will print an INFO log to logs of the JOB at
https://developer.mend.io/github/letsql/letsql