feat: add image and array manipulation functions
This commit demonstrates how to add array and image manipulation functions
to LETSQL, at the same time it shows how to store this data.

The idea to store the data is to use nested list arrays to represent matrices and
binary blobs to represent images.

The functions added are

- segment_anything
- tensor_mean_all
- rotate

Additionally high level operators for segment_anything and rotate
were included