fix: add duckdb and xgboost as dependencies for examples (#216)
Currently, we use dask for tokenizing; dask selects the fastest hashing algorithm available from a list of possible hashing libraries (cityhash, xxhash, murmurhash, etc.). This causes LETSQL not to monkey-patch (hotfix) some attributes due to differences in the hashes. To solve this problem, cityhash was added as a core dependency since it will be the first library selected by dask. Additionally, a warning is shown to the user in case LETSQL cannot import cityhash.  

In addition, a new GitHub action (ci-test-examples) was added to test the examples, using only the examples extras installation.

Finally, duckdb and xgboost were added to the `examples` extras as dependencies.