chore: fix pre-release workflow
What?
The workflow was failing at the build stage for the Linux arch aarch64

Why?
The reason was the new Cargo dependencies for reading a csv from http(s)/s3.
In particular the ring transitive dependency: https://crates.io/crates/ring

Solution:
The solution is to drop the architectures armv7, s390x, and ppc64le, which are not supported by Datafusion or Polars.

Switch the manylinux to `2_28` when the architecture is aarch64 as suggested by:

https://github.com/briansmith/ring/issues/1728#issuecomment-1758180655

This change was also implemented by Datafusion

https://github.com/apache/datafusion-python/blob/fe0738a9c0b536cdf20b0dc0455d14a0d16d2835/.github/workflows/build.yml#L218-L219