chore: snowflake-connector-python security update
Snowflake discovered and remediated a vulnerability in the Snowflake Connector for Python.
A function from the snowflake.connector.pandas_tools module is vulnerable to SQL injection.
This vulnerability affects versions 2.2.5 through 3.13.0. Snowflake fixed the issue in
version 3.13.1.

Vulnerability Details
A function from the snowflake.connector.pandas_tools module is not sanitizing
all of its arguments, and queries using them are not parametrized. An attacker
controlling these arguments could achieve SQL injection by passing crafted input.
Any SQL executed that way by an attacker would still run in the context of the current session.

Solution
Snowflake released version 3.13.1 of the Snowflake Connector for Python,
which fixes this issue. We recommend users upgrade to version 3.13.1.